polaroid/requirements.txt
amqp==1.4.9
aniso8601==3.0.2
anyjson==0.3.3
appnope==0.1.0
backcall==0.1.0
billiard==3.3.0.23
celery==3.1.18
certifi==2018.11.29
decorator==4.3.2
Django==2.1.7
django-celery==3.2.2
django-graphql-jwt==0.2.0
django-redis==4.10.0
django-rest-framework==0.1.0
djangorestframework==3.9.1
graphene==2.1.3
graphene-django==2.2.0
graphql-core==2.1
graphql-relay==0.4.5
ipython==7.2.0
ipython-genutils==0.2.0
jedi==0.13.2
kombu==3.0.37
parso==0.3.4
pexpect==4.6.0
pickleshare==0.7.5
promise==2.2.1
prompt-toolkit==2.0.8
psycopg2==2.7.7
psycopg2-binary==2.7.7
ptyprocess==0.6.0
Pygments==2.3.1
PyJWT==1.7.1
pytz==2018.9
redis==2.10.6
Rx==1.6.1
singledispatch==3.4.0.3
six==1.12.0
traitlets==4.3.2
wcwidth==0.1.7
yapf==0.26.0
